:sparkles: Add shortcut recognition to rename command
rename command will now addition check for a shortcut pattern if the necessary amount of arguments are not found. Shortcuts follow the patterrn 'vn' or 'tn', where n equals the index of the channel you wish to rename (starting from 1) and 'v' or 't' denote the channel type. Rename command can now also be invoked using the shorthand 'rn'.